% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmBackground.R
\name{adjustImg}
\alias{adjustImg}
\title{adjust image}
\usage{
adjustImg(img, offset)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image}}

\item{offset}{offset of the image.}
}
\value{
an object of \link[EBImage:Image-class]{Image}
}
\description{
\link[EBImage:Image-class]{Image} based on output of \link{alignmentImageData}
}
