% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmBackground.R
\name{alignmentImageData}
\alias{alignmentImageData}
\title{alignment the images}
\usage{
alignmentImageData(figureA, figureB)
}
\arguments{
\item{figureA, figureB}{\link[EBImage:Image-class]{Image} object.}
}
\value{
offset of x and y
}
\description{
get offset of two of \link[EBImage:Image-class]{Image}.
Can not handle rotation now.
}
