% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCounter.R
\name{cellCounter}
\alias{cellCounter}
\title{cell counter}
\usage{
cellCounter(
  file,
  channel = "green",
  offset = 0.05,
  cellSizeRange = c(20, 1000),
  distanceSameCell = 10,
  xmlfile = sub("\\\\.(tiff|tif)$", ".cellCounter.xml", file, ignore.case = TRUE),
  imageFilename = sub("\\\\.(tiff|tif)$", ".czi", basename(file), ignore.case = TRUE),
  counterType = c(prenew = 5, new = 4, old = 6),
  zvalue = fixZvalue,
  adjustPipeline = c(GaussianBlur, ScurveAdjust),
  detectFun = detectObjects,
  saveAdjustImage = NULL,
  silence = FALSE,
  ...
)
}
\arguments{
\item{file}{tiff file name}

\item{channel}{channel to be detected}

\item{offset}{the offset of color from background, (0, 1).}

\item{cellSizeRange}{cell size range}

\item{distanceSameCell}{distance of same cell in different frame}

\item{xmlfile}{filename of xml}

\item{imageFilename}{filename of the original czi file}

\item{counterType}{counterType in Cell Counter of imageJ}

\item{zvalue}{change the z to original z.}

\item{adjustPipeline}{adjust pipeline before cell detection}

\item{detectFun}{the function used to detect the cell, see \link{detectObjects}, \link{detectObjects2}.}

\item{saveAdjustImage}{the file name for adjusted image. NULL to ignore saving.}

\item{silence}{output the message or not}

\item{...}{parameters could be used in the pipeline}
}
\description{
count the cells from a z-stack tiff or czi file and output xml for imageJ
}
\examples{
library(EBImage)
library(scales)
library(XML)
##library(rJava)
##.jinit(parameters=c("-Xms1g","-Xmx4g"))
##library(RBioFormats) ## to support czi file,
## try to install by devtools::install_github("aoles/RBioFormats")
cellCounter(system.file("extdata", "sample.tiff", package="cellCounter"),
xmlfile="sample.xml", imageFilename="sample.czi")

}
\author{
Jianhong Ou
}
