% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPipelineFun.R
\name{changeCellSize}
\alias{changeCellSize}
\title{changeCellSize}
\usage{
changeCellSize(
  img,
  changeCellSize_direction = c("erode", "dilate"),
  changeCellSize_targetChannel = c("red", "green", "blue"),
  changeCellSize_size = 3,
  changeCellSize_shape = "disc",
  ...
)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image}.}

\item{changeCellSize_direction}{"erode" or "dilate". erode make the cell size smaller 
and dilate make the cell size bigger.}

\item{changeCellSize_targetChannel}{the target channel.}

\item{changeCellSize_size, changeCellSize_shape}{Brush size and shape, see \link[EBImage:makeBrush]{makeBrush}}

\item{\dots}{not used.}
}
\description{
apply rescale Image data to [0, 1]
}
\examples{
library(EBImage)
img <- readImage(system.file("extdata", "low.jpg", package="cellCounter"))
changeCellSize(img)
}
\author{
Jianhong Ou
}
