% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropImages.R
\name{cropImages}
\alias{cropImages}
\title{crop the images}
\usage{
cropImages(imgsList, positions)
}
\arguments{
\item{imgsList}{output of \link{rmBackground}.}

\item{positions}{a list of positions of x and y for crop. They are lefttop corner and rightbottom corner.}
}
\value{
a list of cropped alignedImage, background and backgroundRemoved.
}
\description{
crop the image by the targeting region
}
