% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectObjects.R
\name{detectObjects}
\alias{detectObjects}
\title{detect objects method1}
\usage{
detectObjects(
  img,
  offset = 0.05,
  objectSizeRange = c(10, 1000),
  size = 31,
  tolerance = 1,
  ext = 1,
  ...
)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image} or \link{Image2}.}

\item{offset}{the offset of color from background, (0, 1).}

\item{objectSizeRange}{numeric(2). object size range in pixel.}

\item{size}{Brush size, see \link[EBImage:makeBrush]{makeBrush}.}

\item{tolerance, ext}{see \link[EBImage:watershed]{watershed}.}

\item{\dots}{not use.}
}
\value{
an label Image.
}
\description{
detect object, good for dense objects.
}
\details{
steps: 
1. use filter2 to creat local background.  
2. the image is convert to 0 or 1 by local background + offset
3. use fillHull to fill the holes
4. use watershed to segment the objects.
5. use bwlabel to label the objects
6. filter the objects by size range.
}
\author{
Jianhong Ou
}
