% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectObjects.R
\name{detectObjects2}
\alias{detectObjects2}
\title{detect objects method2}
\usage{
detectObjects2(img, offset = 0.2, objectSizeRange = c(10, 1000), size = 5,
  w = 40, h = 40, tolerance = 1, ext = 1, ...)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image} or \link{Image2}.}

\item{offset, w, h}{offset,w,h for \link[EBImage:thresh]{thresh}}

\item{objectSizeRange}{numeric(2). object size range in pixel.}

\item{size}{Brush size, see \link[EBImage:makeBrush]{makeBrush}.}

\item{tolerance, ext}{see \link[EBImage:watershed]{watershed}.}

\item{\dots}{not use.}
}
\value{
an label Image.
}
\description{
detect object
}
\details{
steps: 
1. all piexl - mean value, 
2. rescale to increase contrast
detect the flys. 
1. use opening to remove some samll objects 
2. use fillHull to fill the holes.
3. use watershed to segment the object 
4. use bwlabel to label the object 
5. filter the object by size range
}
\author{
Jianhong Ou
}
