% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectObjects.R
\name{detectObjects3}
\alias{detectObjects3}
\title{detect objects method3}
\usage{
detectObjects3(img, offset = 0.05, objectSizeRange = c(10, 1000),
  tolerance = 1, ext = 1, ...)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image} of \link{Image2}.}

\item{offset}{the offset of color from background, (0, 1).}

\item{objectSizeRange}{numeric(2). object size range in pixel.}

\item{tolerance, ext}{see \link[EBImage:watershed]{watershed}.}

\item{\dots}{not use.}
}
\value{
an label Image.
}
\description{
detect object, good for dense objects.
}
\details{
steps: 
1. use watershed to segment the objects.
2. use bwlabel to label the objects
3. filter the objects by size range.
}
\author{
Jianhong Ou
}
