% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterImage.R
\name{filterImage}
\alias{filterImage}
\title{filter image}
\usage{
filterImage(img, ..., expand.size = 21)
}
\arguments{
\item{img}{An object of \link[EBImage:Image-class]{Image} or \link{Image2}}

\item{\dots}{filter criteria}

\item{expand.size}{extend the regions that meet the filter criteria by pixels. 
A numeric containing the size of the brush in pixels. 
This should be an odd number; even numbers are rounded to the next odd one, 
i.e., size = 4 has the same effect as size = 5. Default is 21}
}
\value{
An object of \link[EBImage:Image-class]{Image} or \link{Image2}
}
\description{
filter pixels by conditions
}
\author{
Jianhong Ou
library(EBImage)
img <- readImage(system.file("extdata", "low.jpg", package="cellCounter"))
img <- filterImage(img, red > .3, green > .3, green + .1 > red, green < red + .3)
}
