% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmBackground.R
\name{getBackgroundAfterAlignment}
\alias{getBackgroundAfterAlignment}
\title{get the background after alignment}
\usage{
getBackgroundAfterAlignment(imgs)
}
\arguments{
\item{imgs}{a list of \link[EBImage:Image-class]{Image}.}
}
\value{
an object of \link[EBImage:Image-class]{Image}
}
\description{
Get the background after \link{alignmentImageData}
}
