% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPipelineFun.R
\name{increaseContrast}
\alias{increaseContrast}
\title{increaseContrast}
\usage{
increaseContrast(img, increaseContrast_times = 2, ...)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image}.}

\item{increaseContrast_times}{a numeric vector.}

\item{\dots}{not used.}
}
\description{
increase the Image contrast by multiply the image data by a value.
}
\examples{
library(EBImage)
img <- readImage(system.file("extdata", "low.jpg", package="cellCounter"))
increaseContrast(img)
}
\author{
Jianhong Ou
}
