% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledImage2Cell.R
\name{labeledImage2Cell}
\alias{labeledImage2Cell}
\title{convert the labeled imaged to a list of cell object}
\usage{
labeledImage2Cell(
  imgs,
  colorChannel = c(red = "#FF0000", green = "#00FF00", blue = "#0000FF")
)
}
\arguments{
\item{imgs}{a list of \link[EBImage:Image-class]{Image} or \link{Image2}. It should be output of detectObjects functions.}

\item{colorChannel}{character(1). The color channel for the cell.}
}
\description{
Convert the output of detectObjects functions into a list of cell object
}
\author{
Jianhong Ou
}
