% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputCellTable.R
\name{outputCellTable}
\alias{outputCellTable}
\title{convert list of cell to data.frame}
\usage{
outputCellTable(cells)
}
\arguments{
\item{cells}{list of \link{cell} object}
}
\value{
a data frame
}
\description{
convert list of cell object to a data frame
}
\author{
Jianhong Ou
}
