% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDetections.R
\name{overlayDetections}
\alias{overlayDetections}
\title{overlay detections}
\usage{
overlayDetections(imgs, cells, col = "orange", output = "Image", ...)
}
\arguments{
\item{imgs}{a list of \link[EBImage:Image-class]{Image} or \link{Image2}.}

\item{cells}{a list of \link{cell} objects.}

\item{col}{color of the labels.}

\item{output}{\link[EBImage:Image-class]{Image} or \link{Image2}}

\item{\dots}{the parameters for \link[graphics:text]{text}}
}
\value{
an object of \link[EBImage:Image-class]{Image}, or a list of \link{Image2}.
}
\description{
prepare data for display detections
}
\author{
Jianhong Ou
}
