% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDetections.R
\name{rasterText}
\alias{rasterText}
\title{raster the text in a image}
\usage{
rasterText(img, ...)
}
\arguments{
\item{img}{an \link[EBImage:Image-class]{Image} or \link{Image2}}

\item{\dots}{the parameters for \link[graphics:text]{text}}
}
\value{
an object of \link[EBImage:Image-class]{Image}.
}
\description{
add text labels in a image
}
\author{
Jianhong Ou
}
