% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{Read images}
\usage{
readFile(file, channel = NULL, Image2mode = FALSE)
}
\arguments{
\item{file}{a character vector of file names.}

\item{channel}{the channel to be read in}

\item{Image2mode}{return should be object of Image or Image2?}
}
\value{
an Image or Image2
}
\description{
Read images from files.
}
\examples{
readFile(system.file("extdata", "sample.tiff", package="cellCounter"))

}
\author{
Jianhong Ou, Yanchao Han
}
