% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\name{readListImg}
\alias{readListImg}
\title{read a list of image}
\usage{
readListImg(files, ...)
}
\arguments{
\item{files}{filenames of the images}

\item{...}{parameters Not used.}
}
\value{
a list of \link{Image2}
}
\description{
use reasonable memory to read a list of images
}
