% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPipelineFun.R
\name{removeBackground}
\alias{removeBackground}
\title{removeBackground}
\usage{
removeBackground(img, removeBackground_background, ...)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image}.}

\item{removeBackground_background}{background cut off value.}

\item{\dots}{not used.}
}
\description{
substract the background from Image data. 
If background is not given, it will be the median of the Image.
}
\examples{
library(EBImage)
img <- readImage(system.file("extdata", "low.jpg", package="cellCounter"))
removeBackground(img)
}
\author{
Jianhong Ou
}
