% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPipelineFun.R
\name{rescaleImage}
\alias{rescaleImage}
\title{rescaleImage}
\usage{
rescaleImage(img, ...)
}
\arguments{
\item{img}{an object of \link[EBImage:Image-class]{Image}.}

\item{\dots}{not used.}
}
\description{
apply rescale Image data to [0, 1].
}
\examples{
library(EBImage)
img <- readImage(system.file("extdata", "low.jpg", package="cellCounter"))
rescaleImage(channel(img, "blue"))
}
\author{
Jianhong Ou
}
