% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmBackground.R
\name{rmBackground}
\alias{rmBackground}
\title{remove background}
\usage{
rmBackground(imgs)
}
\arguments{
\item{imgs}{a list of \link[EBImage:Image-class]{Image} or \link{Image2}.}
}
\value{
a list of alignedImage, background and backgroundRemoved.
}
\description{
remove the background based on a batch of image
}
