% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParentOfCells.R
\name{setParentOfCells}
\alias{setParentOfCells}
\title{given a list of cell,
set the parent of offsprings of the cells
depend on their number of frames and their position.
1. distance of same cell must have overlaps
2. same cell must be the nearest one
3. same cell in same frame can have multiple colors, and the other informations should be merge
4.}
\usage{
setParentOfCells()
}
\description{
given a list of cell,
set the parent of offsprings of the cells
depend on their number of frames and their position.
1. distance of same cell must have overlaps
2. same cell must be the nearest one
3. same cell in same frame can have multiple colors, and the other informations should be merge
4.
}
