% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCounterXML.R
\name{writeCounterXML}
\alias{writeCounterXML}
\title{write counters to xml file}
\usage{
writeCounterXML(counterTypeTable, xmlfile, originalImageFileName)
}
\arguments{
\item{counterTypeTable}{counter type table, contain columns x, y, z, type}

\item{xmlfile}{output xml file name}

\item{originalImageFileName}{original image file name}
}
\description{
write counter table to xmlfile
}
\examples{
counterTypeTable = data.frame(x=1, y=1, z=1, type=1)
writeCounterXML(counterTypeTable, tempfile(), "test")
}
\author{
Jianhong Ou
}
