% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{cell-class}
\alias{cell-class}
\alias{cell}
\alias{parent}
\alias{parent,cell-method}
\alias{parent<-}
\alias{parent<-,cell-method}
\alias{offsprings}
\alias{offsprings,cell-method}
\alias{offsprings<-}
\alias{offsprings<-,cell-method}
\alias{$,cell-method}
\title{Class cell}
\usage{
cell(...)

parent(x)

parent(x) <- value

offsprings(x)

offsprings(x) <- value

\S4method{$}{cell}(x, name)
}
\arguments{
\item{\dots}{Each argument in \dots becomes an slot in the new cell.}

\item{x}{an object of cell.}

\item{value}{the value to be applied.}

\item{name}{slot name of cell}
}
\description{
An object of class cell represents a cell
}
\section{Slots}{

\describe{
\item{\code{cx}}{the center position of x}

\item{\code{cy}}{the center position of y}

\item{\code{xs}}{the positions of x in the cell}

\item{\code{ys}}{the positions of y in the cell}

\item{\code{id}}{the id of the cell}

\item{\code{parent}}{the parent of the cell}

\item{\code{offsprings}}{the offsprings of the cell}

\item{\code{frame}}{the frame of the cell}

\item{\code{color}}{the color the cell}
}}

\examples{
cell()

}
\author{
Jianhong Ou
}
