% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{Image2-class}
\alias{Image2-class}
\alias{Image2}
\alias{is.Image2}
\alias{Image2}
\alias{readImage2}
\alias{writeImage2}
\alias{toImage}
\alias{toImage,Image2-method}
\alias{imageData}
\alias{imageData,Image2-method}
\alias{imageData,Image2-method}
\alias{imageData<-}
\alias{imageData<-,Image2-method}
\alias{imageData<-,Image2-method}
\alias{colorMode}
\alias{colorMode,Image2-method}
\alias{colorMode<-}
\alias{colorMode<-,Image2-method}
\alias{channel}
\alias{channel,Image2-method}
\alias{show,Image2-method}
\alias{display}
\alias{display,Image2-method}
\title{Class Image2}
\usage{
is.Image2(x)

Image2(img)

readImage2(file)

writeImage2(x, ...)

toImage(img2, frames)

imageData(y)

\S4method{imageData}{Image2}(y)

imageData(y) <- value

\S4method{imageData}{Image2}(y) <- value

colorMode(y)

colorMode(y) <- value

channel(x, mode)

\S4method{show}{Image2}(object)

display(x, method, frames, ...)
}
\arguments{
\item{x}{an object of \link{Image2}}

\item{img}{an \link[EBImage:Image-class]{Image} object.}

\item{file}{file name of image.}

\item{\dots}{parameters could passed to \link[EBImage:io]{writeImage}.}

\item{img2}{an \link{Image2} object.}

\item{frames}{frames of the images to display}

\item{y}{an \link{Image2} object.}

\item{value}{the image data or colormode.}

\item{mode}{A character value specifying the target mode for conversion. See \link[EBImage:channel]{channel}.}

\item{object}{an \link{Image2} object.}

\item{method}{see \link[EBImage:display]{display}}
}
\description{
An object of class Image2 represents a Image with HDF5Array

Method imageData

Method imageData<-

Method colorMode

Method colorMode<-

Method channel

Method show

Method display
}
\section{Slots}{

\describe{
\item{\code{seed,index,delayed_ops}}{see \link[DelayedArray:DelayedArray-class]{DelayedArray}}

\item{\code{colormode}}{colormode of the image.}
}}

\examples{
library(EBImage)
img <- readImage(system.file("extdata", "low.jpg", package="cellCounter")) 
Image2(img)

}
\author{
Jianhong Ou
}
