% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDetections.R
\name{overlayDetections}
\alias{overlayDetections}
\title{overlay detections}
\usage{
overlayDetections(imgsList)
}
\arguments{
\item{imgsList}{output of \link{cropImages}.}
}
\value{
an object of \link[EBImage:Image-class]{Image}.
}
\description{
prepare data for display detections
}
